Convert util/osm stuff to @id-sdk/util (note utilEntityAndDeepMemberIDs is missing, next sdk update will include it)
