+++ modules/actions/reflect.js
@@ -1,6 +1,5 @@
-import { vecInterp, vecLength } from '@id-sdk/vector';
-import { geomGetSmallestSurroundingRectangle } from '@id-sdk/geom';
-import { utilGetAllNodes } from '../util';
+import { geomGetSmallestSurroundingRectangle, vecInterp, vecLength } from '@id-sdk/math';
+import { utilGetAllNodes } from '@id-sdk/util';
 
 
 /* Reflect the given area around its axis of symmetry */
