+++ modules/actions/rotate.js
@@ -1,9 +1,8 @@
-import { geomRotatePoints } from '@id-sdk/geom';
-import { utilGetAllNodes } from '../util';
+import { geomRotatePoints } from '@id-sdk/math';
+import { utilGetAllNodes } from '@id-sdk/util';
 
 
 export function actionRotate(rotateIds, pivot, angle, projection) {
-
     var action = function(graph) {
         return graph.update(function(graph) {
             utilGetAllNodes(rotateIds, graph).forEach(function(node) {
