+++ modules/core/validator.js
@@ -1,11 +1,11 @@
 import { dispatch as d3_dispatch } from 'd3-dispatch';
 import { Extent } from '@id-sdk/math';
-import { utilArrayChunk, utilArrayGroupBy } from '@id-sdk/util';
+import { utilArrayChunk, utilArrayGroupBy, utilEntityAndDeepMemberIDs } from '@id-sdk/util';
 
 import { prefs } from './preferences';
 import { coreDifference } from './difference';
 import { modeSelect } from '../modes/select';
-import { utilEntityAndDeepMemberIDs, utilRebind } from '../util';
+import { utilRebind } from '../util';
 import * as Validations from '../validations/index';
