+++ modules/operations/disconnect.js
@@ -1,10 +1,10 @@
-import { utilArrayUniq } from '@id-sdk/util';
+import { utilArrayUniq, utilGetAllNodes } from '@id-sdk/util';
 
 import { t } from '../core/localizer';
 import { actionDisconnect } from '../actions/disconnect';
 import { behaviorOperation } from '../behavior/operation';
 import { prefs } from '../core/preferences';
-import { utilGetAllNodes, utilTotalExtent } from '../util/util';
+import { utilTotalExtent } from '../util/util';
 
 
 export function operationDisconnect(context, selectedIDs) {
