+++ modules/operations/rotate.js
@@ -1,8 +1,10 @@
+import { utilGetAllNodes } from '@id-sdk/util';
+
 import { t } from '../core/localizer';
 import { behaviorOperation } from '../behavior/operation';
 import { modeRotate } from '../modes/rotate';
 import { prefs } from '../core/preferences';
-import { utilGetAllNodes, utilTotalExtent } from '../util/util';
+import { utilTotalExtent } from '../util/util';
 
 
 export function operationRotate(context, selectedIDs) {
