+++ modules/ui/raw_tag_editor.js
@@ -1,6 +1,6 @@
 import { dispatch as d3_dispatch } from 'd3-dispatch';
 import { select as d3_select } from 'd3-selection';
-import { utilArrayDifference } from '@id-sdk/util';
+import { utilArrayDifference, utilTagDiff } from '@id-sdk/util';
 
 import { t } from '../core/localizer';
 import { services } from '../services';
@@ -8,7 +8,7 @@ import { svgIcon } from '../svg/icon';
 import { uiCombobox } from './combobox';
 import { uiDisclosure } from './disclosure';
 import { uiTagReference } from './tag_reference';
-import { utilGetSetValue, utilNoAuto, utilRebind, utilTagDiff } from '../util';
+import { utilGetSetValue, utilNoAuto, utilRebind } from '../util';
 
 
 export function uiRawTagEditor(context) {
