+++ modules/ui/sections/raw_tag_editor.js
@@ -1,6 +1,6 @@
 import { dispatch as d3_dispatch } from 'd3-dispatch';
 import { select as d3_select } from 'd3-selection';
-import { utilArrayDifference, utilArrayIdentical } from '@id-sdk/util';
+import { utilArrayDifference, utilArrayIdentical, utilTagDiff } from '@id-sdk/util';
 
 import { services } from '../../services';
 import { svgIcon } from '../../svg/icon';
@@ -9,7 +9,7 @@ import { uiSection } from '../section';
 import { uiTagReference } from '../tag_reference';
 import { prefs } from '../../core/preferences';
 import { t } from '../../core/localizer';
-import { utilGetSetValue, utilNoAuto, utilRebind, utilTagDiff } from '../../util';
+import { utilGetSetValue, utilNoAuto, utilRebind } from '../../util';
 
 
 export function uiSectionRawTagEditor(id, context) {
