+++ modules/util/index.js
@@ -1,20 +1,11 @@
 export { utilAsyncMap } from './util';
-export { utilCleanTags } from './clean_tags';
-export { utilCombinedTags } from './util';
-export { utilDeepMemberSelector } from './util';
 export { utilDetect } from './detect';
 export { utilDisplayName } from './util';
 export { utilDisplayNameForPath } from './util';
 export { utilDisplayType } from './util';
 export { utilDisplayLabel } from './util';
-export { utilEntityRoot } from './util';
-export { utilEntityAndDeepMemberIDs } from './util';
-export { utilEntityOrMemberSelector } from './util';
-export { utilEntityOrDeepMemberSelector } from './util';
-export { utilEntitySelector } from './util';
 export { utilFastMouse } from './util';
 export { utilFunctor } from './util';
-export { utilGetAllNodes } from './util';
 export { utilGetSetValue } from './get_set_value';
 export { utilHighlightEntities } from './util';
 export { utilKeybinding } from './keybinding';
@@ -23,8 +14,6 @@ export { utilPrefixCSSProperty } from './util';
 export { utilPrefixDOMProperty } from './util';
 export { utilRebind } from './rebind';
 export { utilSetTransform } from './util';
-export { utilTagDiff } from './util';
-export { utilTagText } from './util';
 export { utilTotalExtent } from './util';
 export { utilTriggerEvent } from './trigger_event';
 export { utilWrap } from './util';
