+++ modules/validations/private_data.js
@@ -1,6 +1,8 @@
+import { utilTagDiff } from '@id-sdk/util';
+
 import { actionChangeTags } from '../actions/change_tags';
 import { t } from '../core/localizer';
-import { utilDisplayLabel, utilTagDiff } from '../util';
+import { utilDisplayLabel } from '../util';
 import { validationIssue, validationIssueFix } from '../core/validation';
