+++ test/index.html
@@ -150,7 +150,6 @@
       'spec/ui/fields/localized.js',
       'spec/ui/fields/wikipedia.js',
 
-      'spec/util/clean_tags.js',
       'spec/util/keybinding.js',
       'spec/util/util.js',
